some changes ?!

+++ stuff/profiler.hh
@@ -242,7 +242,7 @@ long Profiler::Output( CollectiveCommunication& comm, InfoContainer& run_infos )
             {
                 TimingData data = it->second;
                 long clock_count = data.end - data.start;
-                clock_count =  factor * long ( comm.sum( clock_count ) / double( CLOCKS_PER_SEC*0.001*numProce ) );
+				clock_count =  long( factor * long ( comm.sum( clock_count ) / double( CLOCKS_PER_SEC*0.001*numProce ) ) );
                 csv << clock_count << "," ;
             }
             csv << factor  << std::endl;
@@ -279,7 +279,7 @@ long Profiler::Output( CollectiveCommunication& comm, InfoContainer& run_infos )
             {
                 TimingData data = it->second;
                 long clock_count = data.end - data.start;
-                clock_count =  factor * long ( comm.sum( clock_count ) / double( CLOCKS_PER_SEC*0.001*numProce ) );
+				clock_count =  long( factor * long ( comm.sum( clock_count ) / double( CLOCKS_PER_SEC*0.001*numProce ) ) );
                 csv << clock_count << "," ;
             }
             csv << factor << std::endl;
